Add register page styling
